refactor: move Rust tests to tests/ directory with proper lib.rs crate structure
Create src-tauri/src/lib.rs as library entry point with all module declarations. Main.rs is now a thin wrapper. Move capabilities tests from src/ to tests/capabilities.rs as integration tests. Follows same pattern as TypeScript tests directory.
