+++ src-tauri/src/capabilities/mod.rs
@@ -6,8 +6,5 @@ pub mod mongo;
 pub mod registry;
 pub mod types;
 
-#[cfg(test)]
-pub mod tests;
-
 pub use registry::registry;
 pub use types::Capability;
